# net/netip: optimize AddrPort.String() method