Improve the efficiency of AddrPort.String() by refactoring the
method to use a shared buffer and reducing memory allocations.
Performance benchmarks show significant improvements in speed
and reduction in bytes per operation, especially for IPv6 addresses.

Fixes #63605

The benchstat result:

```
benchstat old.out new.out 
goos: darwin
goarch: amd64
pkg: net/netip
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
                                   │    old.out    │               new.out               │
                                   │    sec/op     │   sec/op     vs base                │
AddrPortString/v4-12                  49.24n ±  1%   50.65n ± 1%   +2.87% (p=0.000 n=10)
AddrPortString/v6-12                  165.9n ±  1%   113.7n ± 3%  -31.43% (p=0.000 n=10)
AddrPortString/v6_ellipsis-12         160.5n ±  1%   113.0n ± 2%  -29.56% (p=0.000 n=10)
AddrPortString/v6_v4-12              119.00n ±  3%   57.52n ± 4%  -51.66% (p=0.000 n=10)
AddrPortString/v6_zone-12             164.0n ±  1%   111.8n ± 1%  -31.86% (p=0.000 n=10)
AddrPortMarshalText/v4-12             51.52n ±  3%   51.02n ± 0%        ~ (p=0.159 n=10)
AddrPortMarshalText/v6-12             117.6n ±  3%   116.2n ± 1%   -1.19% (p=0.012 n=10)
AddrPortMarshalText/v6_ellipsis-12    119.5n ±  3%   118.0n ± 1%        ~ (p=0.085 n=10)
AddrPortMarshalText/v6_v4-12          63.55n ± 14%   63.25n ± 2%        ~ (p=0.739 n=10)
AddrPortMarshalText/v6_zone-12        115.7n ±  1%   117.4n ± 3%   +1.47% (p=0.003 n=10)
geomean                               103.1n         85.85n       -16.77%

                                   │  old.out   │               new.out                │
                                   │    B/op    │    B/op     vs base                  │
AddrPortString/v4-12                 24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortString/v6-12                 96.00 ± 0%   48.00 ± 0%  -50.00% (p=0.000 n=10)
AddrPortString/v6_ellipsis-12        56.00 ± 0%   32.00 ± 0%  -42.86% (p=0.000 n=10)
AddrPortString/v6_v4-12              56.00 ± 0%   32.00 ± 0%  -42.86% (p=0.000 n=10)
AddrPortString/v6_zone-12            56.00 ± 0%   32.00 ± 0%  -42.86% (p=0.000 n=10)
AddrPortMarshalText/v4-12            24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6-12            64.00 ± 0%   64.00 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6_ellipsis-12   64.00 ± 0%   64.00 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6_v4-12         64.00 ± 0%   64.00 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6_zone-12       64.00 ± 0%   64.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                              52.63        41.51       -21.12%
¹ all samples are equal

                                   │  old.out   │               new.out                │
                                   │ allocs/op  │ allocs/op   vs base                  │
AddrPortString/v4-12                 1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortString/v6-12                 2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
AddrPortString/v6_ellipsis-12        2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
AddrPortString/v6_v4-12              2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
AddrPortString/v6_zone-12            2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)
AddrPortMarshalText/v4-12            1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6-12            1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6_ellipsis-12   1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6_v4-12         1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
AddrPortMarshalText/v6_zone-12       1.000 ± 0%   1.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                              1.320        1.000       -24.21%
¹ all samples are equal
```