@aimuz 2024-01-30
Repeat fix, close https://go-review.googlesource.com/c/go/+/557777